Fixed calibration update bug.
Originally, we removed updating the calibration from the dictionary iteration in _updateSensor and updated it later so that we would get the dictionary form of the sensor outside of the loop. However, this wasn't working right, since it referenced value, which was defined in the dictionary iteration.
This version places calibration update in the iteration, but causes a lookup. May revert to previous method later.